[US_ND] Map IC violations to `REVOCATION` (Recidiviz/recidiviz-data#42430)
## Description of the change

I learned yesterday that the `TA_TYPE` 17 is used when a person was on
supervision in North Dakota, having been sentenced in another state, and
is sent back to their sentencing state after a violation. They then go
through revocation proceedings in the sending state. This remaps those
instances to `REVOCATION` end reasons instead of `DISCHARGE`.

This brings the number of incarceration starts with an inflow from
supervision that have a `NEW_ADMISSION` start reason down by around 20,
in service of https://github.com/Recidiviz/recidiviz-data/issues/33540

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3d80c8379f64c760af35344cfc96c09114996b7e