[Justice Counts] Delete JC Dockerfile (Recidiviz/recidiviz-data#42795)
## Description of the change

Removes JC Dockerfile. The JC app is no longer deployed to staging or
production and as a result the Docker image is out of date and causing
CI test failures.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b84ee1737e3ca99df75bbae56564052a2d9bed96