[Email Reminders] Send email reminders for PA, minor bugfixes (Recidiviz/recidiviz-data#42855)
## Description of the change

Four small changes in this PR:
- Add Pennsylvania to supervision line staff email reminders. Everyone
was trained by the end of April. This copy was approved by Kirtana,
Thomas and Fy.
- Squash a small bug that came up during the email send today where
`isSupervisors` wasn't defined due to the refactor to use `userType`.
- Change the number of sheets that `CheckLoginStatus` covers to 3, to
include the new facilities login reminders sheet. Also add a prompt to
hide the previous month's sheet tabs, which I think will make it easier
to manage the spreadsheet of sent emails.
- Fix an off-by-one in the denominator reported by `CheckLoginStatus`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#41579

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 058a44ea627950ec14ae92dc5136619096f3cd08