[testing] add db-backed tests for `InvalidateOperationsDBFilesController` (Recidiviz/recidiviz-data#42947)
## Description of the change

this pr updates the tests for `InvalidateOperationsDBFilesController` to
use the db-backed testing paradigm. i was a little worried about the
correctness of the join logic for Recidiviz/recidiviz-data#42040; in preparation for that, i'm
switching our existing tests over to use a db instead. in order to do
so, i made `InvalidateOperationsDBFilesController` inherit from the
`CloudSqlConnectionMixin` and also added an option to spin up the cloud
sql mixin from w/in the mixin

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for Recidiviz/recidiviz-data#42040

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 36ebf475538f1becff8ccff8c23d17a3e65cc800