[US_UT] Hydrate state supervision violation type (Recidiviz/recidiviz-data#42993)
## Description of the change

This hydrates the `StateSupervisionViolationTypeEntry` field on state
supervision violations. I did my best to categorize these correctly -
one thing that I found odd is that violations that stem from controlled
substance use or possession are not labeled as new crimes in any way in
the raw data, and there are not charges that relate to the violation. I
mapped those to `TECHNICAL` violations because of that.

[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-05-28_12_51_14-3492739794711048824;graphView=1?project=recidiviz-staging&inv=1&invt=AbyoLA&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fd76cf2a78e8beb0e1935004437e5a214627eb6d