[US_UT] Add complexity to supervision type mapping (Recidiviz/recidiviz-data#43036)
## Description of the change

We were previously only mapping `supervision_period.supervision_type`
based on a person's legal status; since those ever denote whether a
person is on `PROBATION` or `PAROLE`, we were not mapping any absconsion
supervision types at ingest-time.

This changes the way we map `supervision_type` to map the type to
`ABSCONSION` for all periods that have a start reason that will be
mapped to `ABSCONSION` (`ABSCOND`, `WALKAWAY`, `ESCAPE`, `AWOL`).

This allows us to observe far more `ABSCONSION` sessions than we could
before:
```
  SELECT
    compartment_level_1,
    compartment_level_2,
    COUNT(*)
  FROM recidiviz-staging.elise_type_sessions.compartment_sessions_materialized
  WHERE state_code= 'US_UT'
    AND compartment_level_2 = 'ABSCONSION'
  GROUP BY 1,2
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cb0bdb92895ebc50b960065a2fe1fd0e1263c16b