[Resource Search] Update Resource Source DB to Postgres13 (Recidiviz/recidiviz-data#43045)
This PR updates the postgres version for the Resource Search Cloud SQL
PostgreSQL instance in the postgresql module to use Postgres13.

The previous value db-custom-1-3840 caused the following error during
terraform apply because it was referencing Postgres16, which in
compatible with the team:

```
Invalid Tier (db-custom-1-3840) for (ENTERPRISE_PLUS) Edition. Use a predefined Tier like db-perf-optimized-N-* instead.
To resolve this, we’ve replaced it with db-perf-optimized-N-2, which is a valid performance-optimized tier for ENTERPRISE_PLUS editions.
```
No other changes are included in this PR.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7201d4b9f7bef2ea21892c8d709f0dc354ae6efe