Revert "[bigquery] remove row level permissions application from view… (Recidiviz/recidiviz-data#43074)
… materialization (Recidiviz/recidiviz-data#42896)"

This reverts commit caa6814fb76e48246775a6f95b2e50cf11686fba.

removing this step from view materialization is causing us to try to
apply permissions to too many tables in the `apply_row_access_policies`
task and we're encountering some underlying throttling issue. this, in
some cases, is causing only a subset of the policy list to be applied,
causing unexpected permission denials when retrieving data.

i think it would be good to drop row level permissions (with retry +
backoff in case this request also fails) if row level permission
application is unsuccessful to avoid partial policy list applications. i
can also try to find a more precise estimate of when we are encountering
these failures and tune concurrency accordingly.

context:
https://recidiviz.slack.com/archives/C028X32LRH7/p1748559577271769

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: feaef492809af7675400d167798c2c8bc326829c