[US_AZ] Added Maybe Eligible logic into Admin Supervision (Recidiviz/recidiviz-data#43016)
## Description of the change

This change implements the Maybe Eligible logic we have discussed
extensively for Admin Supervision.
The change is made by creating an almost eligible feature for the three
multi-pronged criteria (MPC), which will eventually incorporate the
single-pronged (AKA normal) criteria once we decide upon which of those
fit into our almost eligible needs.
In the opportunity record query, we create three new columns for the
MPC™ to signify when a resident is missing at least one, at most all of
them.

I also fixed a typo in the TPR/DTP TES files where the
EligibleCriteriaCondition was incorrectly stating that a resident is
*missing* that criteria when in reality it means they *meet* that
criteria.

Uploaded to sandbox under `shalin`.

We will need to make changes in the admin panel to show these criteria
in our dynamic requirements to check tab.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#42480

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d3783b68da4bf63b9259c097a3a66fa3cab24904