[Email reminders] Update email text for linestaff to show opportunity-specific numbers (Recidiviz/recidiviz-data#41916)
## Description of the change

Updates the text of the email reminders for line staff who have not
logged in. Instead of displaying one bullet point that summarizes the
number of available opportunities, it now displays a bullet point with
the number of eligible and/or almost eligible opportunities for each
opportunity type. The total number of opportunities is still displayed
in a paragraph below the bullet point list.

This is dependent on Recidiviz/recidiviz-data#42853, which I don't think will be in prod until
6/10, but I've tested the query with sandbox data at
`etiry_separate_eligible`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#41347

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 230d752c34d940a4b277b2ae419843fab2e6baae