[BigQuery] Apply admin row-level access policy first (Recidiviz/recidiviz-data#43313)
## Description of the change

When we generate row-level access policy queries, the queries are
currently generated (and therefore applied) in the following order:
1) Drop all row-level policies
2) (for each relevant state) Create access policy for this state
3) For all states without restrictive policies, create full access
policy for those states
4) Create policy which grants access to all rows for admins (i.e.
service accounts)

If we fail between 2 and 4, we'll find ourselves in a situation where a
service account might only have access to data for some states. This PR
restructures so the admin policy comes first, meaning we now have this
order:
1) Drop all row-level policies
2) Create policy which grants access to all rows for admins (i.e.
service accounts)
3) (for each relevant state) Create access policy for this state
4) For all states without restrictive policies, create full access
policy for those states

If we fail to apply 1, we'll still have all the old policies, so no
issue. If we fail between 1 and 2, there will be no row-level policies
on the table which is not good for security but IS good for the service
account / admins being able to access data. If we fail after 2, service
accounts will still have full access.

This PR also includes some minor changes to make the queries always
generate in a deterministic order.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See https://recidiviz.slack.com/archives/C028X32LRH7/p1749228682299789

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f533e36b443e385aeb88153d8a968b0cf231af87