[Resource Search] Turn On Enable Point In Time Recovery When Backups Are Enabled (Recidiviz/recidiviz-data#43283)
Point in time recovery must be disabled when backup is disabled. I've
updated the variable to depend on the `is_backup_enabled` flag so that
always reflects if backups are enabled/disabled.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9a92d837309788d92ce6292b63e3b1ee0787166f