[US_TN] Update custodial authority override to only look at adjacent periods (Recidiviz/recidiviz-data#42830)
## Description of the change

Updates custodial authority normalization in TN to only overwrite
custodial authority for -T periods with the custodial authority from a
preceding -P period if the -P period is adjacent (has a release_date
equal to the admission date of the -T period). A T period is still
considered adjacent to a P period if there are other T periods in
between the two, and all the periods are contiguous.

That is, if someone has the following periods:

1. P period, custodial authority = COUNTY, t=1 to t=2
2. T period, custodial authority = STATE_PRISON, t=2 to t=3
3. T period, custodial authority = STATE_PRISON, t=3 to t=4

Both T periods will be normalized with custodial_authority COUNTY,
because period (2) is adjacent to period (1), and period (3) is adjacent
to period (2).

**Similarly, if someone has the following periods:**
1. P period, custodial authority = COUNTY, t=1 to t=2
2. T period, custodial authority = COUNTY, t=2 to t=3
3. T period, custodial authority = STATE_PRISON, t=3 to t=4

The result will be the same. The third period will still be normalized
with custodial_authority COUNTY, because period (2) is adjacent to
period (1), and period (3) is adjacent to period (2). Even though period
(2) itself doesn't have its custodial authority changed, it's still
taken into account when determining contiguity between periods 1 and 3.

Also adds a T flag to the CUST_CHANGE and UNIT_MOVEMENT movement
reasons, so that periods with these admission reasons don't break up the
chain of contiguity between a P period and a T period.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cd7e0ee4ee24242464ae2520bccec5f8c674204a