[Resource Search] Initial Migration for DBs (Recidiviz/recidiviz-data#43374)
This PR adds the initial alembic migration for the new RESOURCE_SEARCH
application. I've tested locally and confirmed that the migration
successfully created the resource, resource_score, and scrape_source
tables.

**These changes were approved and merged previously, but were reverted
because the cloudsql database did not yet exist in recidiviz-staging and
caused many errors. I used terraform to create cloudsql DB this morning
so now this change is safe to merge. **

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 159aeca4652d141608bfa8cb67c57e0df3db019a